src/backend/pyproject.toml
[project]
name = "framer-backend"
version = "0.1.0"
description = "Framer - AI-assisted pre-development thinking framework"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "gitpython>=3.1.40",
    "aiosqlite>=0.19.0",
    "pyyaml>=6.0.1",
    "markdown-it-py>=3.0.0",
    "httpx>=0.26.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
]
ai = [
    "agno>=0.1.0",
    "openai>=1.10.0",
    "anthropic>=0.18.0",
]
vector = [
    "chromadb>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["../../tests/backend"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
